+++ projects/ng-translation/src/lib/models/index.ts
@@ -3,6 +3,7 @@ export * from './locale.model';
 export * from './resource.model';
 export * from './resource-list.model';
 export * from './resource-loader.model';
+export * from './translatable-language-list.model';
 export * from './translatable-option.model';
 export * from './translatable-option-list.model';
 export * from './translatable-text-list.model';